Move Network class and use removesuffix
Improvements from Franz in PR
